(frontend) Update the device configuration feature, list of devices is refreshing and visual display of the form is correct
This update is in order to fix the display of the list when a device is updated.
It misses form fields verification and props are not yet updating in form fields configuration.
